Integrated from gtnh-flow master Wonderful.
