+++ src/gtnhvelo/graph/_postProcessing.py
@@ -1,25 +1,167 @@
 import logging
 import math
-import re
 from collections import defaultdict
 from copy import deepcopy
+from string import ascii_uppercase
 
+from gtnhvelo.data.basicTypes import Ingredient, IngredientCollection, Recipe
 from gtnhvelo.graph._utils import _iterateOverMachines
 from gtnhvelo.gtnh.overclocks import OverclockHandler
 
+from typing import TYPE_CHECKING
+if TYPE_CHECKING:
+    from gtnhvelo.graph import Graph
 
-def _addPowerLineNodes(self):
-    # This checks for burnables being put into sink and converts them to EU/t
+
+def capitalizeMachine(machine):
+    # check if machine has capitals, and if so, preserve them
+    capitals = set(ascii_uppercase)
+    machine_capitals = [ch for ch in machine if ch in capitals]
+
+    capitalization_exceptions = {
+        # Format is old_str: new_str
+    }
+
+    if len(machine_capitals) > 0:
+        return machine
+    elif machine in capitalization_exceptions:
+        return capitalization_exceptions[machine]
+    else:
+        return machine.title()
+
+
+def createMachineLabels(self):
+    # Distillation Tower
+    # ->
+    # 5.71x HV Distillation Tower
+    # Cycle: 2.0s
+    # Amoritized: 1.46K EU/t
+    # Per Machine: 256EU/t
+
+    overclock_data = self.parent_context.data['overclock_data']
+
+    for node_id in self.nodes:
+        if self._checkIfMachine(node_id):
+            rec_id = node_id
+            rec = self.recipes[rec_id]
+        else:
+            continue
+
+        label_lines = []
+
+        # Standard label
+
+        default_label = [
+            f'{round(rec.multiplier, 2)}x {rec.user_voltage.upper()} {capitalizeMachine(rec.machine)}',
+            f'Cycle: {rec.dur/20}s',
+            f'Amoritized: {self.userRound(int(round(rec.eut, 0)))} EU/t',
+            f'Per Machine: {self.userRound(int(round(rec.base_eut, 0)))} EU/t',
+        ]
+
+        label_lines.extend(default_label)
+
+        if self.graph_config['POWER_UNITS'] != 'eut':
+            if self.graph_config['POWER_UNITS'] == 'auto':
+                tier_idx = overclock_data['voltage_data']['tiers'].index(rec.user_voltage)
+            else:
+                tier_idx = overclock_data['voltage_data']['tiers'].index(self.graph_config['POWER_UNITS'])
+            voltage_at_tier = self.tierToVoltage(tier_idx)
+            label_lines[-2] = f'Amoritized: {self.userRound(int(round(rec.eut, 0)) / voltage_at_tier)} {overclock_data["voltage_data"]["tiers"][tier_idx].upper()}'
+            label_lines[-1] = f'Per Machine: {self.userRound(int(round(rec.base_eut, 0)) / voltage_at_tier)} {overclock_data["voltage_data"]["tiers"][tier_idx].upper()}'
+
+        # Edits for power machines
+        recognized_basic_power_machines = {
+            # "basic" here means "doesn't cost energy to run"
+            'gas turbine',
+            'combustion gen',
+            'semifluid gen',
+            'steam turbine',
+            'rocket engine',
+
+            'large naquadah reactor',
+            'large gas turbine',
+            'large steam turbine',
+            'large combustion engine',
+            'extreme combustion engine',
+            'XL Turbo Gas Turbine',
+            'XL Turbo Steam Turbine',
+
+            'air intake hatch',
+        }
+        if rec.machine in recognized_basic_power_machines:
+            # Remove power input data
+            label_lines = label_lines[:-2]
+
+        line_if_attr_exists = {
+            'heat': (lambda rec: f'Base Heat: {rec.heat}K'),
+            'coils': (lambda rec: f'Coils: {rec.coils.title()}'),
+            'saw_type': (lambda rec: f'Saw Type: {rec.saw_type.title()}'),
+            'material': (lambda rec: f'Turbine Material: {rec.material.title()}'),
+            'size': (lambda rec: f'Size: {rec.size.title()}'),
+            'efficiency': (lambda rec: f'Efficiency: {rec.efficiency}'),
+            'wasted_fuel': (lambda rec: f'Wasted Fuel: {rec.wasted_fuel}'),
+        }
+        for lookup, line_generator in line_if_attr_exists.items():
+            if hasattr(rec, lookup):
+                label_lines.append(line_generator(rec))
+
+        self.nodes[rec_id]['label'] = '\n'.join(label_lines)
+
+
+def addUserNodeColor(self):
+    targeted_nodes = [i for i, x in self.recipes.items() if getattr(x, 'target', False) != False]
+    numbered_nodes = [i for i, x in self.recipes.items() if getattr(x, 'number', False) != False]
+    all_user_nodes = set(targeted_nodes) | set(numbered_nodes)
+
+    for rec_id in all_user_nodes:
+        self.nodes[rec_id].update({'fillcolor': self.graph_config['LOCKEDNODE_COLOR']})
+
+
+def addMachineMultipliers(self):
+    # Compute machine multiplier based on solved ingredient quantities
+    # FIXME: If multipliers disagree, sympy solver might have failed on an earlier step
+
+    for rec_id, rec in self.recipes.items():
+        multipliers = []
+
+        for io_dir in ['I', 'O']:
+            for ing in getattr(rec, io_dir):
+                ing_name = ing.name
+                base_quant = ing.quant
+
+                # Look up edge value from sympy solver
+                solved_quant_per_s = 0
+                for edge in self.adj[rec_id][io_dir]:
+                    if edge[2] == ing_name:
+                        # print(edge, self.edges[edge]['quant'])
+                        solved_quant_per_s += self.edges[edge]['quant']
+
+                base_quant_s = base_quant / (rec.dur / 20)
+
+                # print(io_dir, rec_id, ing_name, getattr(rec, io_dir))
+                # print(solved_quant_per_s, base_quant_s, rec.dur)
+                # print()
+
+                machine_multiplier = solved_quant_per_s / base_quant_s
+                multipliers.append(machine_multiplier)
+
+        final_multiplier = max(multipliers)
+        rec.multiplier = final_multiplier
+        rec.eut = rec.multiplier * rec.eut
+
+
+def addPowerLineNodesV2(self: 'Graph'):
     generator_names = {
         0: 'gas turbine',
         1: 'combustion gen',
         2: 'semifluid gen',
         3: 'steam turbine',
-        4: 'rocket engine fuel',
+        4: 'rocket engine',
         5: 'large naquadah reactor',
     }
 
     power_data = self.parent_context.data['power_data']
+    overclock_data = self.parent_context.data['overclock_data']
 
     turbineables = power_data['turbine_fuels']
     combustables = power_data['combustion_fuels']
@@ -34,103 +176,129 @@ def _addPowerLineNodes(self):
     known_burnables.update({x: [4, y] for x, y in rocket_fuels.items()})
     known_burnables.update({x: [5, y] for x, y in naqline_fuels.items()})
 
+    # Add new burn machines to graph - they will be computed for using new solver
+    # 1. Find highest voltage on the chart - use this for burn generator tier
+    # 2. Figure out highest node index on the chart - use this for adding generator nodes
+    voltages = overclock_data['voltage_data']['tiers']
+    highest_voltage = 0
+    highest_node_index = 0
+    for rec_id, rec in self.recipes.items():
+
+        rec_voltage = voltages.index(rec.user_voltage)
+        if rec_voltage > highest_voltage:
+            highest_voltage = rec_voltage
+
+        int_index = int(rec_id)
+        if int_index > highest_node_index:
+            highest_node_index = int_index
+
+    highest_node_index += 1
+
+    # 3. Redirect burnables currently going to sink and redirect them to a new burn machine
     outputs = self.adj['sink']['I']
-    generator_number = 1
     for edge in deepcopy(outputs):
         node_from, _, ing_name = edge
         edge_data = self.edges[edge]
-        quant = edge_data['quant']
+        quant_s = edge_data['quant']
 
         if ing_name in known_burnables and not ing_name in self.graph_config['DO_NOT_BURN']:
             self.parent_context.cLog(f'Detected burnable: {ing_name.title()}! Adding to chart.', level=logging.INFO)
             generator_idx, eut_per_cell = known_burnables[ing_name]
-            gen_name = generator_names[generator_idx].title()
+            gen_name = generator_names[generator_idx]
 
             # Add node
-            node_gen = f'power_{generator_number}_{generator_idx}'
-            generator_number += 1
-            node_name = f'{gen_name} (100% eff)'
+            node_idx = f'{highest_node_index}'
+
+            # Burn gen is a singleblock
+            def findClosestVoltage(voltage_list, voltage):
+                nonlocal voltages
+                leftmost = voltages.index(voltage_list[0])
+                rightmost = voltages.index(voltage_list[-1])
+                target = voltages.index(voltage)
+
+                # First try to voltage down
+                if rightmost < target:
+                    return voltages[rightmost]
+                elif leftmost <= target <= rightmost:
+                    return voltages[target]
+                elif leftmost > target:
+                    return voltages[leftmost]
+
+            available_efficiencies = power_data['simple_generator_efficiencies'][gen_name]
+            gen_voltage = findClosestVoltage(list(available_efficiencies), voltages[highest_voltage])
+            efficiency = available_efficiencies[gen_voltage]
+
+            # Compute I/O for a single tick
+            gen_voltage_index = voltages.index(gen_voltage)
+            output_eut = 32 * (4 ** gen_voltage_index)
+            loss_on_singleblock_output = (2 ** (gen_voltage_index + 1))
+            expended_eut = output_eut + loss_on_singleblock_output
+
+            expended_fuel_t = expended_eut / (eut_per_cell / 1000 * efficiency)
+
+            gen_input = IngredientCollection(
+                Ingredient(
+                    ing_name,
+                    expended_fuel_t
+                )
+            )
+            gen_output = IngredientCollection(
+                Ingredient(
+                    'EU',
+                    output_eut
+                )
+            )
+
+            # Append to recipes
+            self.recipes[str(highest_node_index)] = Recipe(
+                gen_name,
+                gen_voltage,
+                gen_input,
+                gen_output,
+                0,
+                1,
+                efficiency=f'{efficiency*100}%',
+                wasted_fuel=f'{self.userRound(loss_on_singleblock_output)}EU/t/amp',
+            )
+
+            produced_eut_s = quant_s / expended_fuel_t * output_eut
+            self.parent_context.cLog(
+                ''.join([
+                    f'Added {gen_voltage} generator burning {quant_s} {ing_name} for '
+                    f'{self.userRound(produced_eut_s/20)}EU/t at {output_eut}EU/t each.'
+                ])
+            )
+
             self.addNode(
-                node_gen,
-                label=node_name,
+                node_idx,
                 fillcolor=self.graph_config['NONLOCKEDNODE_COLOR'],
                 shape='box'
             )
 
             # Fix edges to point at said node
-            produced_eut = eut_per_cell * quant / 1000
             # Edge (old output) -> (generator)
             self.addEdge(
                 node_from,
-                node_gen,
+                node_idx,
                 ing_name,
-                quant,
+                quant_s,
                 **edge_data['kwargs'],
             )
             # Edge (generator) -> (EU sink)
             self.addEdge(
-                node_gen,
+                node_idx,
                 'sink',
                 'EU',
-                produced_eut,
+                produced_eut_s,
             )
             # Remove old edge and repopulate adjacency list
             del self.edges[edge]
             self.createAdjacencyList()
 
-    # Automatically balance the outputs of UCFE
-    # 1. Get UCFE node
-    UCFE_id = None
-    for rec_id, rec in self.recipes.items():
-        if rec.machine == 'universal chemical fuel engine':
-            UCFE_id = rec_id
-
-    if UCFE_id is not None:
-        self.parent_context.cLog('Detected UCFE, autobalancing...', level=logging.INFO)
-
-        # 2. Determine whether non-combustion promoter input is combustable or gas
-        input_ingredient_collection = self.recipes[UCFE_id].I
-        if len(input_ingredient_collection) != 2:
-            raise RuntimeError('Too many or too few inputs to UCFE - expected 2.')
-
-        if 'combustion promoter' not in input_ingredient_collection._ingdict:
-            raise RuntimeError('UCFE detected, but "combustion promoter" is not one of its inputs. Cannot autobalance.')
-
-        fuel_name = ''
-        for ing in input_ingredient_collection._ings:
-            if ing.name != 'combustion promoter':
-                fuel_name = ing.name
-                break
-
-        burn_value_table = None
-        if fuel_name in turbineables:
-            burn_value_table = turbineables
-            coefficient = 0.04
-        elif fuel_name in combustables:
-            burn_value_table = combustables
-            coefficient = 0.04
-        elif fuel_name in rocket_fuels:
-            burn_value_table = rocket_fuels
-            coefficient = 0.005
-        else:
-            raise RuntimeError(
-                f'Unrecognized input fuel to UCFE: {fuel_name}. Can only burn gas, combustables, or rocket fuel.')
-
-        # 3. Compute UCFE ratio and output EU/s
-        combustion_promoter_quant = input_ingredient_collection['combustion promoter'][0]
-        fuel_quant = input_ingredient_collection[fuel_name][0]
-        ratio = combustion_promoter_quant / fuel_quant
-        self.parent_context.cLog(f'UCFE power ratio: {ratio}', level=logging.INFO)
-
-        efficiency = math.exp(-coefficient * ratio) * 1.5
-        self.parent_context.cLog(f'Efficiency stat: {efficiency}', level=logging.INFO)
-        output_eu = efficiency * burn_value_table[fuel_name] * (fuel_quant / 1000)
-
-        # 4. Update edge with new value
-        self.edges[(UCFE_id, 'sink', 'EU')]['quant'] = output_eu
+            highest_node_index += 1
 
 
-def _addSummaryNode(self):
+def addSummaryNode(self):
     # Now that tree is fully locked, add I/O node
     # Specifically, inputs are adj[source] and outputs are adj[sink]
     misc_data = self.parent_context.data['special_machine_weights']
