+++ src/gtnhvelo/graph/_solver.py
@@ -12,9 +12,25 @@ from sympy.solvers import solve
 from sympy.sets.sets import EmptySet
 
 from gtnhvelo.data.basicTypes import Ingredient, IngredientCollection, Recipe
-from gtnhvelo.graph import Graph
 
 from rich.progress import Progress
+
+
+from gtnhvelo.graph import Graph
+from gtnhvelo.graph._preProcessing import (
+    connectGraph,
+    removeBackEdges,
+)
+from gtnhvelo.graph._postProcessing import (
+    addMachineMultipliers,
+    addPowerLineNodesV2,
+    addSummaryNode,
+    addUserNodeColor,
+    createMachineLabels,
+)
+from gtnhvelo.graph._output import (
+    outputGraphviz
+)
 
 
 class SympySolver:
@@ -447,64 +463,62 @@ class SympySolver:
                 self.graph.parent_context.cLog(
                     f'Between output={self.graph.recipes[mpdm_cache[0][0]].O}', level=logging.WARNING)
                 self.graph.parent_context.cLog(
-                    f'and input={self.graph.recipes[mpdm_cache[1][0]].I}', level=logging.WARNING)
+                    f'    and  input={self.graph.recipes[mpdm_cache[1][0]].I}', level=logging.WARNING)
 
-                # TODO: Fix this
-                # Probably will take a while
+                self.graph.parent_context.cLog('Please fix by either:', level=logging.INFO)
 
-                # self.graph.parent_context.cLog('Please fix by either:', level=logging.INFO)
+                # if constant_diff < 0:
+                #     parent_group_idx = 0
+                #     child_group_idx = 1
+                # else:
+                parent_group_idx = 0
+                child_group_idx = 1
 
-                # # if constant_diff < 0:
-                # #     parent_group_idx = 0
-                # #     child_group_idx = 1
-                # # else:
-                # parent_group_idx = 0
-                # child_group_idx = 1
+                # Negative means too much of right side, or too few of other sided inputs
+                self.graph.parent_context.cLog(
+                    f'1. Sending excess {group[parent_group_idx]} {product} to sink', level=logging.INFO)
 
-                # # Negative means too much of right side, or too few of other sided inputs
-                # self.graph.parent_context.cLog(f'1. Sending excess {group[parent_group_idx]} {product} to sink', level=logging.INFO)
+                # Check other sided inputs
+                machine, product, direction, multi_idx = idx_to_mpdm[var_to_idx(group[child_group_idx])]
+                nonself_product = []
+                for edge in self.graph.adj[machine][direction]:
+                    # print(self.graph.adj[machine])
+                    # print(edge)
+                    a, b, edgeproduct = edge
+                    if edgeproduct != product:
+                        nonself_product.append((
+                            edgeproduct,
+                            'v' + f'{self.edge_from_perspective_to_index[(edge, machine)]}',
+                        ))
 
-                # # Check other sided inputs
-                # machine, product, direction, multi_idx = idx_to_mpdm[var_to_idx(group[child_group_idx])]
-                # nonself_product = []
-                # for edge in self.graph.adj[machine][direction]:
-                #     # print(self.graph.adj[machine])
-                #     # print(edge)
-                #     a, b, edgeproduct = edge
-                #     if edgeproduct != product:
-                #         nonself_product.append((
-                #             edgeproduct,
-                #             'v' + f'{self.edge_from_perspective_to_index[(edge, machine)]}',
-                #         ))
-
-                # self.graph.parent_context.cLog(f'2. Pulling more {nonself_product} from source', level=logging.INFO)
+                self.graph.parent_context.cLog(f'2. Pulling more {nonself_product} from source', level=logging.INFO)
 
                 # Output graph for end user to view
                 self._debugAddVarsToEdges()
-                self.graph.outputGraphviz()
+                outputGraphviz(self.graph)
 
-                # # TODO: Automate solution process fully
+                # TODO: Automate solution process fully
 
-                # selection = input()  # TODO: Verify input
+                selection = input()  # TODO: Verify input
 
-                # if selection == '1':
-                #     # Send excess to sink
-                #     # 1. Similar to multi-IO: (a-c could probably be spun off into another fxn)
-                #     #       a. reassociate old variable with machine sum of product
-                #     #       b. create a new variable for old edge
-                #     #       c. create a new variable for machine -> sink
-                #     # 2. Redo linear solve
-                #     # 3. Give option for user to add new I/O association to YAML config (will delete comments)
-                #     pass
-                # elif selection == '2':
-                #     # Pull more of each other input from source
-                #     # 1. Similar to multi-IO: (a-c could probably be spun off into another fxn)
-                #     #       a. reassociate each old variable on all sides of machine with machine sum of product
-                #     #       b. create a new variable for each old edge
-                #     #       c. create a new variable for each source -> machine
-                #     # 2. Redo linear solve
-                #     # 3. Give option for user to add new I/O association to YAML config (will delete comments)
-                #     pass
+                if selection == '1':
+                    # Send excess to sink
+                    # 1. Similar to multi-IO: (a-c could probably be spun off into another fxn)
+                    #       a. reassociate old variable with machine sum of product
+                    #       b. create a new variable for old edge
+                    #       c. create a new variable for machine -> sink
+                    # 2. Redo linear solve
+                    # 3. Give option for user to add new I/O association to YAML config (will delete comments)
+                    pass
+                elif selection == '2':
+                    # Pull more of each other input from source
+                    # 1. Similar to multi-IO: (a-c could probably be spun off into another fxn)
+                    #       a. reassociate each old variable on all sides of machine with machine sum of product
+                    #       b. create a new variable for each old edge
+                    #       c. create a new variable for each source -> machine
+                    # 2. Redo linear solve
+                    # 3. Give option for user to add new I/O association to YAML config (will delete comments)
+                    pass
 
     def _debugAddVarsToEdges(self):
         # Add variable indices to edges and rec_id to machines
@@ -569,8 +583,6 @@ class SympySolver:
                     solution_index = self.edge_from_perspective_to_index[(edge, a)]
                 elif b_machine:
                     solution_index = self.edge_from_perspective_to_index[(edge, b)]
-                else:
-                    raise NotImplementedError('How did this happen?')
 
                 quant = self.solved_vars[solution_index]
                 relevant_edge = self.graph.edges[edge]
@@ -578,296 +590,10 @@ class SympySolver:
                 relevant_edge['locked'] = True  # TODO: Legacy - check if can be removed
 
 
-def addMachineMultipliers(self):
-    # Compute machine multiplier based on solved ingredient quantities
-    # FIXME: If multipliers disagree, sympy solver might have failed on an earlier step
-
-    for rec_id, rec in self.recipes.items():
-        multipliers = []
-
-        for io_dir in ['I', 'O']:
-            for ing in getattr(rec, io_dir):
-                ing_name = ing.name
-                base_quant = ing.quant
-
-                # Look up edge value from sympy solver
-                solved_quant_per_s = 0
-                for edge in self.adj[rec_id][io_dir]:
-                    if edge[2] == ing_name:
-                        # print(edge, self.edges[edge]['quant'])
-                        solved_quant_per_s += self.edges[edge]['quant']
-
-                base_quant_s = base_quant / (rec.dur / 20)
-
-                # print(io_dir, rec_id, ing_name, getattr(rec, io_dir))
-                # print(solved_quant_per_s, base_quant_s, rec.dur)
-                # print()
-
-                machine_multiplier = solved_quant_per_s / base_quant_s
-                multipliers.append(machine_multiplier)
-
-        final_multiplier = max(multipliers)
-        rec.multiplier = final_multiplier
-        rec.eut = rec.multiplier * rec.eut
-
-
-def capitalizeMachine(machine):
-    # check if machine has capitals, and if so, preserve them
-    capitals = set(ascii_uppercase)
-    machine_capitals = [ch for ch in machine if ch in capitals]
-
-    capitalization_exceptions = {
-
-    }
-
-    if len(machine_capitals) > 0:
-        return machine
-    elif machine in capitalization_exceptions:
-        return capitalization_exceptions[machine]
-    else:
-        return machine.title()
-
-
-def createMachineLabels(self):
-    # Distillation Tower
-    # ->
-    # 5.71x HV Distillation Tower
-    # Cycle: 2.0s
-    # Amoritized: 1.46K EU/t
-    # Per Machine: 256EU/t
-
-    overclock_data = self.parent_context.data['overclock_data']
-
-    for node_id in self.nodes:
-        if self._checkIfMachine(node_id):
-            rec_id = node_id
-            rec = self.recipes[rec_id]
-        else:
-            continue
-
-        label_lines = []
-
-        # Standard label
-
-        default_label = [
-            f'{round(rec.multiplier, 2)}x {rec.user_voltage.upper()} {capitalizeMachine(rec.machine)}',
-            f'Cycle: {rec.dur/20}s',
-            f'Amoritized: {self.userRound(int(round(rec.eut, 0)))} EU/t',
-            f'Per Machine: {self.userRound(int(round(rec.base_eut, 0)))} EU/t',
-        ]
-
-        label_lines.extend(default_label)
-
-        if self.graph_config['POWER_UNITS'] != 'eut':
-            if self.graph_config['POWER_UNITS'] == 'auto':
-                tier_idx = overclock_data['voltage_data']['tiers'].index(rec.user_voltage)
-            else:
-                tier_idx = overclock_data['voltage_data']['tiers'].index(self.graph_config['POWER_UNITS'])
-            voltage_at_tier = self.tierToVoltage(tier_idx)
-            label_lines[-2] = f'Amoritized: {self.userRound(int(round(rec.eut, 0)) / voltage_at_tier)} {overclock_data["voltage_data"]["tiers"][tier_idx].upper()}'
-            label_lines[-1] = f'Per Machine: {self.userRound(int(round(rec.base_eut, 0)) / voltage_at_tier)} {overclock_data["voltage_data"]["tiers"][tier_idx].upper()}'
-
-        # Edits for power machines
-        recognized_basic_power_machines = {
-            # "basic" here means "doesn't cost energy to run"
-            'gas turbine',
-            'combustion gen',
-            'semifluid gen',
-            'steam turbine',
-            'rocket engine',
-
-            'large naquadah reactor',
-            'large gas turbine',
-            'large steam turbine',
-            'large combustion engine',
-            'extreme combustion engine',
-            'XL Turbo Gas Turbine',
-            'XL Turbo Steam Turbine',
-
-            'air intake hatch',
-        }
-        if rec.machine in recognized_basic_power_machines:
-            # Remove power input data
-            label_lines = label_lines[:-2]
-
-        line_if_attr_exists = {
-            'heat': (lambda rec: f'Base Heat: {rec.heat}K'),
-            'coils': (lambda rec: f'Coils: {rec.coils.title()}'),
-            'saw_type': (lambda rec: f'Saw Type: {rec.saw_type.title()}'),
-            'material': (lambda rec: f'Turbine Material: {rec.material.title()}'),
-            'size': (lambda rec: f'Size: {rec.size.title()}'),
-            'efficiency': (lambda rec: f'Efficiency: {rec.efficiency}'),
-            'wasted_fuel': (lambda rec: f'Wasted Fuel: {rec.wasted_fuel}'),
-        }
-        for lookup, line_generator in line_if_attr_exists.items():
-            if hasattr(rec, lookup):
-                label_lines.append(line_generator(rec))
-
-        self.nodes[rec_id]['label'] = '\n'.join(label_lines)
-
-
-def addPowerLineNodesV2(self):
-    generator_names = {
-        0: 'gas turbine',
-        1: 'combustion gen',
-        2: 'semifluid gen',
-        3: 'steam turbine',
-        4: 'rocket engine',
-        5: 'large naquadah reactor',
-    }
-
-    power_data = self.parent_context.data['power_data']
-    overclock_data = self.parent_context.data['overclock_data']
-
-    turbineables = power_data['turbine_fuels']
-    combustables = power_data['combustion_fuels']
-    semifluids = power_data['semifluids']
-    rocket_fuels = power_data['rocket_fuels']
-    naqline_fuels = power_data['naqline_fuels']
-
-    known_burnables = {x: [0, y] for x, y in turbineables.items()}
-    known_burnables.update({x: [1, y] for x, y in combustables.items()})
-    known_burnables.update({x: [2, y] for x, y in semifluids.items()})
-    known_burnables['steam'] = [3, 500]
-    known_burnables.update({x: [4, y] for x, y in rocket_fuels.items()})
-    known_burnables.update({x: [5, y] for x, y in naqline_fuels.items()})
-
-    # Add new burn machines to graph - they will be computed for using new solver
-    # 1. Find highest voltage on the chart - use this for burn generator tier
-    # 2. Figure out highest node index on the chart - use this for adding generator nodes
-    voltages = overclock_data['voltage_data']['tiers']
-    highest_voltage = 0
-    highest_node_index = 0
-    for rec_id, rec in self.recipes.items():
-
-        rec_voltage = voltages.index(rec.user_voltage)
-        if rec_voltage > highest_voltage:
-            highest_voltage = rec_voltage
-
-        int_index = int(rec_id)
-        if int_index > highest_node_index:
-            highest_node_index = int_index
-
-    highest_node_index += 1
-
-    # 3. Redirect burnables currently going to sink and redirect them to a new burn machine
-    outputs = self.adj['sink']['I']
-    for edge in deepcopy(outputs):
-        node_from, _, ing_name = edge
-        edge_data = self.edges[edge]
-        quant_s = edge_data['quant']
-
-        if ing_name in known_burnables and not ing_name in self.graph_config['DO_NOT_BURN']:
-            self.parent_context.cLog(f'Detected burnable: {ing_name.title()}! Adding to chart.', level=logging.INFO)
-            generator_idx, eut_per_cell = known_burnables[ing_name]
-            gen_name = generator_names[generator_idx]
-
-            # Add node
-            node_idx = f'{highest_node_index}'
-
-            # Burn gen is a singleblock
-            def findClosestVoltage(voltage_list, voltage):
-                nonlocal voltages
-                leftmost = voltages.index(voltage_list[0])
-                rightmost = voltages.index(voltage_list[-1])
-                target = voltages.index(voltage)
-
-                # First try to voltage down
-                if rightmost < target:
-                    return voltages[rightmost]
-                elif leftmost <= target <= rightmost:
-                    return voltages[target]
-                elif leftmost > target:
-                    return voltages[leftmost]
-
-            available_efficiencies = power_data['simple_generator_efficiencies'][gen_name]
-            gen_voltage = findClosestVoltage(list(available_efficiencies), voltages[highest_voltage])
-            efficiency = available_efficiencies[gen_voltage]
-
-            # Compute I/O for a single tick
-            gen_voltage_index = voltages.index(gen_voltage)
-            output_eut = 32 * (4 ** gen_voltage_index)
-            loss_on_singleblock_output = (2 ** (gen_voltage_index + 1))
-            expended_eut = output_eut + loss_on_singleblock_output
-
-            expended_fuel_t = expended_eut / (eut_per_cell / 1000 * efficiency)
-
-            gen_input = IngredientCollection(
-                Ingredient(
-                    ing_name,
-                    expended_fuel_t
-                )
-            )
-            gen_output = IngredientCollection(
-                Ingredient(
-                    'EU',
-                    output_eut
-                )
-            )
-
-            # Append to recipes
-            self.recipes[str(highest_node_index)] = Recipe(
-                gen_name,
-                gen_voltage,
-                gen_input,
-                gen_output,
-                0,
-                1,
-                efficiency=f'{efficiency*100}%',
-                wasted_fuel=f'{self.userRound(loss_on_singleblock_output)}EU/t/amp',
-            )
-
-            produced_eut_s = quant_s / expended_fuel_t * output_eut
-            self.parent_context.cLog(
-                ''.join([
-                    f'Added {gen_voltage} generator burning {quant_s} {ing_name} for '
-                    f'{self.userRound(produced_eut_s/20)}EU/t at {output_eut}EU/t each.'
-                ]),
-                level=logging.INFO,
-            )
-
-            self.addNode(
-                node_idx,
-                fillcolor=self.graph_config['NONLOCKEDNODE_COLOR'],
-                shape='box'
-            )
-
-            # Fix edges to point at said node
-            # Edge (old output) -> (generator)
-            self.addEdge(
-                node_from,
-                node_idx,
-                ing_name,
-                quant_s,
-                **edge_data['kwargs'],
-            )
-            # Edge (generator) -> (EU sink)
-            self.addEdge(
-                node_idx,
-                'sink',
-                'EU',
-                produced_eut_s,
-            )
-            # Remove old edge and repopulate adjacency list
-            del self.edges[edge]
-            self.createAdjacencyList()
-
-            highest_node_index += 1
-
-
-def addUserNodeColor(self):
-    targeted_nodes = [i for i, x in self.recipes.items() if getattr(x, 'target', False) != False]
-    numbered_nodes = [i for i, x in self.recipes.items() if getattr(x, 'number', False) != False]
-    all_user_nodes = set(targeted_nodes) | set(numbered_nodes)
-
-    for rec_id in all_user_nodes:
-        self.nodes[rec_id].update({'fillcolor': self.graph_config['LOCKEDNODE_COLOR']})
-
-
 def graphPreProcessing(self, progress_cb):
-    self.connectGraph()
+    connectGraph(self)
     progress_cb(6.6)
-    self.removeBackEdges()
+    removeBackEdges(self)
     progress_cb(6.6)
     self.createAdjacencyList()
     progress_cb(6.6)
@@ -881,9 +607,8 @@ def graphPostProcessing(self, progress_cb):
     progress_cb(6.6)
     createMachineLabels(self)
     progress_cb(6.6)
-
-    self._addSummaryNode()
-
+    addSummaryNode(self)
+    progress_cb(6.6)
     addUserNodeColor(self)
     progress_cb(6.6)
 
@@ -898,15 +623,16 @@ def systemOfEquationsSolverGraphGen(self, project_name, recipes, graph_config, t
     with Progress(disable=False if not self.quiet else True, transient=True) as progress:
         task = progress.add_task(f'[cyan]{project_name}', total=100)
         def update_progress(advance: float): return progress.update(task, advance=advance)
-
         g = Graph(project_name, recipes, self, graph_config=graph_config, title=title)
-        graphPreProcessing(g, progress_cb=update_progress)
+        self._graph = g  # For test access
+
+        graphPreProcessing(g, update_progress)
 
         g.parent_context.cLog('Running linear solver...', level=logging.INFO)
         solver = SympySolver(g)
-        solver.run(progress_cb=update_progress)
+        solver.run(update_progress)
 
-        graphPostProcessing(g, progress_cb=update_progress)
-        g.outputGraphviz()
+        graphPostProcessing(g, update_progress)
+        outputGraphviz(g)
         update_progress(100)
         progress.stop()
