+++ tests/sanity_config.yaml
@@ -1,20 +1,23 @@
+# gtnhvelo CONFIGURATION FILE
+# DELETE THIS FILE TO REGENERATE UPON RUNNING
+CONFIG_VER: 0
+
+GRAPHVIZ: path
+# Input your graphviz /Graphivz/bin directory path in quotes
+# Type 'path' if you want to use the PATH environment variable
+
 # === FUNCTIONALITY SETTINGS ===
 POWER_LINE: true # Automatically burns all leftover fuels
 DO_NOT_BURN: # If POWER_LINE is true, don't burn these
-  - toluene
-  - phenol
   - hydrogen
-  - creosote oil
-  - sulfuric naphtha
-  - sulfuric gas
-  - sulfuric heavy fuel
-  - sulfuric light fuel
   - coal tar oil
+  - phenol
   - naphtha
-  - creosote oil
   - biomass
   - oil
-OUTPUT_FORMAT: png # png or pdf (pdf is searchable)
+  - steam
+  - creosote oil
+OUTPUT_FORMAT: svg # png or pdf (pdf is searchable)
 USE_RAINBOW_EDGES: true # feature to help distinguish between overlapping edges
 DUR_FORMAT: sec # choose either "ticks" or "sec"
 VIEW_ON_COMPLETION: false
@@ -25,46 +28,70 @@ BOTTLENECK_MIN_VOLTAGE: IV
 MAX_BOTTLENECKS: 5
 
 # === VISUALS ===
-GENERAL_FONT: arial  # nodes and edges
-SUMMARY_FONT: arial
-GROUP_FONT: verdana
-NODE_FONTSIZE: 11
-EDGE_FONTSIZE: 10
-GROUP_FONTSIZE: 18
-BACKGROUND_COLOR: '#043742'
+## === FONTS ===
+GENERAL_FONT: noto sans # nodes and edges
+SUMMARY_FONT: noto sans
+GROUP_FONT: renogare
+TITLE_FONT: renogare
+
+NODE_FONTSIZE: 14
+EDGE_FONTSIZE: 13
+GROUP_FONTSIZE: 20
+TITLE_FONTSIZE: 40
+
+## === COLORS ===
+BACKGROUND_COLOR: '#140e52'
+TITLE_COLOR: '#ffffff'
+SUMMARY_COLOR: '#ffffff'
 EDGECOLOR_CYCLE:
-  - '#b58900' # 'yellow'
-  - '#cb4b16' # 'orange'
-  - '#dc322f' # 'red'
-  - '#d33682' # 'magenta'
-  - '#6c71c4' # 'violet'
-  - '#268bd2' # 'blue'
-  - '#2aa198' # 'cyan'
-  - '#859900' # 'green'
-SOURCESINK_COLOR: 'ghostwhite'
-NONLOCKEDNODE_COLOR: 'lightblue2'
-LOCKEDNODE_COLOR: 'green'
-POSITIVE_COLOR: '#859900'
-NEGATIVE_COLOR: '#dc322f'
+  # Legacy colors
+  # - '#b58900' # 'yellow'
+  # - '#cb4b16' # 'orange'
+  # - '#dc322f' # 'red'
+  # - '#d33682' # 'magenta'
+  # - '#6c71c4' # 'violet'
+  # - '#268bd2' # 'blue'
+  # - '#2aa198' # 'cyan'
+  # - '#859900' # 'green'
+  - '#ff6961'
+  - '#ffb480'
+  - '#f8f38d'
+  - '#42d6a4'
+  - '#08cad1'
+  - '#59adf6'
+  - '#9d94ff'
+  - '#c780e8'
+SOURCESINK_COLOR: 'white'
+NONLOCKEDNODE_COLOR: 'lightcyan'
+LOCKEDNODE_COLOR: 'chartreuse'
+POSITIVE_COLOR: 'chartreuse'
+NEGATIVE_COLOR: 'crimson'
+
+## === EXTRA ===
+POWER_UNITS: 'eut'
+# Only affects nodes
+# eut: Use absolute eu/t numbers, example: 32 EU/t
+# auto: Use user-defined power tier
+# lv, mv, hv, ev, etc.: Use amps
 
 ORIENTATION: 'TB'
 # TB: Vertical, LR: Horizontal, BT: Vertical Inverted, RL: Horizontal Inverted
 # [TB, LR, BT, RL]
 
 LINE_STYLE: 'spline'
-# [line, spline, polyline, curved] 
+# [line, spline, polyline, curved]
 
-RANKSEP: '1.25'
-NODESEP: '0.25'
-# Node separation size in X/Y. 
+RANKSEP: '1.5'
+NODESEP: '0.5'
+# Node separation size in X/Y.
 # This is a suggestion, the graphing algorithm will ignore numbers that are too small.
 
 COMBINE_INPUTS: true
 COMBINE_OUTPUTS: true
 # Whether to (visually) create a new meta node before multi-input/after multi-output ports
 
-
 # { You should probably only modify these if you are dev
+DEBUG_LOGGING: false
 DEBUG_SHOW_EVERY_STEP: false
 SHOW_MACHINE_INDICES: false
 STRIP_BRACKETS: true
