--- tests/test_output/loopGraph
@@ -1,14 +1,0 @@
-digraph {
-	graph [bgcolor="#140e52" fontcolor="#ffffff" fontname=renogare fontsize=40 labelloc=t nodesep=0.5 rankdir=TB ranksep=1.5 splines=spline]
-	source [label=<<table border="0" cellspacing="0"><tr><td><table border="0" cellspacing="0"><tr><td border="0">source</td></tr></table></td></tr><tr><td><table border="0" cellspacing="0"><tr><td border="1" PORT="o_acetic_acid">Acetic Acid</td><td border="1" PORT="o_sulfuric_acid">Sulfuric Acid</td></tr></table></td></tr></table>> fillcolor=white fontname="noto sans" fontsize=14 shape=plain style=filled]
-	sink [label=<<table border="0" cellspacing="0"><tr><td><table border="0" cellspacing="0"><tr><td border="1" PORT="i_water">Water</td><td border="1" PORT="i_ethenone">Ethenone</td><td border="1" PORT="i_sulfuric_acid">Sulfuric Acid</td></tr></table></td></tr><tr><td><table border="0" cellspacing="0"><tr><td border="0">sink</td></tr></table></td></tr></table>> fillcolor=white fontname="noto sans" fontsize=14 shape=plain style=filled]
-	0 [label=<<table border="0" cellspacing="0"><tr><td><table border="0" cellspacing="0"><tr><td border="1" PORT="i_diluted_sulfuric_acid">Diluted Sulfuric Acid</td></tr></table></td></tr><tr><td><table border="0" cellspacing="0"><tr><td border="0">1.0x MV Distillation Tower<br />Cycle: 30.0s<br />Amoritized: 120 EU/t<br />Per Machine: 120 EU/t</td></tr></table></td></tr><tr><td><table border="0" cellspacing="0"><tr><td border="1" PORT="o_sulfuric_acid">Sulfuric Acid</td><td border="1" PORT="o_water">Water</td></tr></table></td></tr></table>> fillcolor=chartreuse fontname="noto sans" fontsize=14 shape=plain style=filled]
-	1 [label=<<table border="0" cellspacing="0"><tr><td><table border="0" cellspacing="0"><tr><td border="1" PORT="i_sulfuric_acid">Sulfuric Acid</td><td border="1" PORT="i_acetic_acid">Acetic Acid</td></tr></table></td></tr><tr><td><table border="0" cellspacing="0"><tr><td border="0">0.8x MV Large Chemical Reactor<br />Cycle: 8.0s<br />Amoritized: 96 EU/t<br />Per Machine: 120 EU/t</td></tr></table></td></tr><tr><td><table border="0" cellspacing="0"><tr><td border="1" PORT="o_ethenone">Ethenone</td><td border="1" PORT="o_diluted_sulfuric_acid">Diluted Sulfuric Acid</td></tr></table></td></tr></table>> fillcolor=lightcyan fontname="noto sans" fontsize=14 shape=plain style=filled]
-	total_io_node [label=<<table border="0"><tr><td align="left"><font color="white" face="noto sans"><b>Summary</b></font></td></tr><hr/><tr><td align="left"><font color="#ff6961" face="noto sans">█ Acetic Acid</font></td><td align ="right"><font color="crimson" face="noto sans">-100.0/s</font></td></tr><tr><td align="left"><font color="#ffb480" face="noto sans">█ Sulfuric Acid</font></td><td align ="right"><font color="crimson" face="noto sans">-33.33/s</font></td></tr><tr><td align="left"><font color="#f8f38d" face="noto sans">█ Water</font></td><td align ="right"><font color="chartreuse" face="noto sans">33.33/s</font></td></tr><tr><td align="left"><font color="#42d6a4" face="noto sans">█ Ethenone</font></td><td align ="right"><font color="chartreuse" face="noto sans">100.0/s</font></td></tr><hr/><tr><td align="left"><font color="white" face="noto sans">Input EU/t:</font></td><td align ="right"><font color="crimson" face="noto sans">-216</font></td></tr><tr><td align="left"><font color="white" face="noto sans">Total machine count:</font></td><td align ="right"><font color="chartreuse" face="noto sans">10.8</font></td></tr><tr><td align="left"><font color="white" face="noto sans">Peak power draw:</font></td><td align ="right"><font color="crimson" face="noto sans">1.88A MV</font></td></tr></table>> color="#ffffff" fillcolor="#140e52" fontname="noto sans" fontsize=14 shape=box style=filled]
-	1:o_diluted_sulfuric_acid:s -> 0:i_diluted_sulfuric_acid:n [arrowhead=normal arrowtail=tee color="#08cad1" dir=both fontcolor="#0cffff" fontname="noto sans" fontsize=13 headlabel="(100.0/s)" labelangle=60 labeldistance=2.5 penwidth=1 taillabel="(100.0/s)"]
-	0:o_water:s -> sink:i_water:n [arrowhead=normal arrowtail=tee color="#f8f38d" dir=both fontcolor="#ffffd3" fontname="noto sans" fontsize=13 headlabel="(33.33/s)" labelangle=60 labeldistance=2.5 penwidth=1 taillabel="(33.33/s)"]
-	source:o_acetic_acid:s -> 1:i_acetic_acid:n [arrowhead=normal arrowtail=tee color="#ff6961" dir=both fontcolor="#ff9d91" fontname="noto sans" fontsize=13 headlabel="(100.0/s)" labelangle=60 labeldistance=2.5 penwidth=1 taillabel="(100.0/s)"]
-	1:o_ethenone:s -> sink:i_ethenone:n [arrowhead=normal arrowtail=tee color="#42d6a4" dir=both fontcolor="#63fff6" fontname="noto sans" fontsize=13 headlabel="(100.0/s)" labelangle=60 labeldistance=2.5 penwidth=1 taillabel="(100.0/s)"]
-	0:o_sulfuric_acid:s -> sink:i_sulfuric_acid:n [arrowhead=normal arrowtail=tee color="#ffb480" dir=both fontcolor="#ffffc0" fontname="noto sans" fontsize=13 headlabel="(66.67/s)" labelangle=60 labeldistance=2.5 penwidth=1 taillabel="(66.67/s)"]
-	source:o_sulfuric_acid:s -> 1:i_sulfuric_acid:n [arrowhead=normal arrowtail=tee color="#ffb480" dir=both fontcolor="#ffffc0" fontname="noto sans" fontsize=13 headlabel="(100.0/s)" labelangle=60 labeldistance=2.5 penwidth=1 taillabel="(100.0/s)"]
-}
